requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1
prefect==2.16.5
prometheus-client==0.19.0
pytesseract==0.3.10
pdfplumber==0.10.4
pikepdf==9.2.0
pillow==10.3.0
pyjwt==2.8.0 